requirements.txt
pyrogram
toml
typer
loguru
rich
peewee
appdirs
pyrubrum-continued==0.2.22
uvloop
python-box
thefuzz[speedup]
python-dateutil
emoji
tgcrypto
aenum